# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: YYC³餐饮行业智能化平台 API
  description: |
    YYC³餐饮行业智能化平台 RESTful API 文档

    ## 平台概述
    YYC³是一个专业的餐饮行业智能化管理平台，提供从点餐、厨房管理、会员管理到营销活动的全方位解决方案。

    ## 主要功能
    - 🍽️ 菜单管理：菜品管理、分类管理、推荐管理
    - 📋 订单管理：订单处理、状态跟踪、统计分析
    - 👥 会员管理：会员信息、等级积分、消费记录
    - 🎯 营销活动：优惠券、促销活动、数据分析
    - 💳 支付管理：多种支付方式、交易记录、退款处理
    - 🏪 门店管理：多门店支持、员工管理、运营分析
    - 📊 报表分析：销售报表、财务报表、运营报表

    ## 认证方式
    本API使用JWT Bearer Token进行身份认证。请在请求头中添加：
    ```
    Authorization: Bearer <your_token>
    ```

    ## 环境信息
    - 生产环境: https://api.yyc3.com
    - 测试环境: https://api-test.yyc3.com
    - 开发环境: http://localhost:3006

  version: 1.0.0
  contact:
    name: YYC³技术支持
    email: support@yyc3.com
    url: https://www.yyc3.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.yyc3.com/v1
    description: 生产环境
  - url: https://api-test.yyc3.com/v1
    description: 测试环境
  - url: http://localhost:3006/v1
    description: 开发环境
...
paths:
...
  /users:
    get:
      tags: [用户管理]
      summary: 获取用户列表
      description: 分页获取系统用户列表
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - name: role
          in: query
          description: 按角色筛选
          schema:
            type: string
            enum: [super_admin, admin, manager, staff]
        - name: status
          in: query
          description: 按状态筛选
          schema:
            type: string
            enum: [active, inactive, locked]
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsersResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [用户管理]
      summary: 创建用户
      description: 创建新的系统用户
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /orders:
    get:
      tags: [订单管理]
      summary: 获取订单列表
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          description: 订单状态
          schema:
            type: string
            enum: [pending, confirmed, preparing, ready, serving, completed, cancelled, refunded]
        - name: storeId
          in: query
          description: 门店ID
          schema:
            type: integer
        - name: startDate
          in: query
          description: 开始日期
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: 结束日期
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedOrdersResponse'

    post:
      tags: [订单管理]
      summary: 创建订单
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /orders/{id}:
    get:
      tags: [订单管理]
      summary: 获取订单详情
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

    patch:
      tags: [订单管理]
      summary: 更新订单状态
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [pending, confirmed, preparing, ready, serving, completed, cancelled]
                reason:
                  type: string
                  description: 状态变更原因
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

  # 会员管理接口
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer Token认证
  parameters:
    Page:
      name: page
      in: query
      description: 页码，从1开始
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: 每页数量
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Search:
      name: search
      in: query
      description: 搜索关键词
      schema:
        type: string
        maxLength: 100
...
    OrderId:
      name: id
      in: path
      description: 订单ID
      required: true
      schema:
        type: integer
        minimum: 1
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "操作成功"
        data:
          type: object
          description: 响应数据
        timestamp:
          type: string
          format: date-time
          description: 响应时间
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: 错误代码
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: 错误信息
              example: "请求参数验证失败"
            statusCode:
              type: integer
              description: HTTP状态码
              example: 400
            details:
              type: object
              description: 错误详情
            timestamp:
              type: string
              format: date-time
              description: 错误时间

    # 分页响应格式
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: 当前页码
          example: 1
        limit:
          type: integer
          description: 每页数量
          example: 20
        total:
          type: integer
          description: 总记录数
          example: 100
        totalPages:
          type: integer
          description: 总页数
          example: 5

    # 认证相关
...
    User:
      type: object
      properties:
        id:
          type: integer
          description: 用户ID
          example: 1
        username:
          type: string
          description: 用户名
          example: "admin"
        email:
          type: string
          description: 邮箱
          example: "admin@yyc3.com"
        fullName:
          type: string
          description: 全名
          example: "系统管理员"
        phone:
          type: string
          description: 手机号
          example: "13800138000"
        avatar:
          type: string
          description: 头像URL
          example: "https://example.com/avatar.jpg"
        role:
          type: string
          enum: [super_admin, admin, manager, staff]
          description: 用户角色
          example: "admin"
        status:
          type: string
          enum: [active, inactive, locked]
          description: 用户状态
          example: "active"
        lastLoginAt:
          type: string
          format: date-time
          description: 最后登录时间
        createdAt:
          type: string
          format: date-time
          description: 创建时间
    CreateUserRequest:
      type: object
      required: [username, email, password, role]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          description: 用户名
        email:
          type: string
          format: email
          description: 邮箱
        password:
          type: string
          minLength: 6
          maxLength: 100
          description: 密码
        fullName:
          type: string
          maxLength: 100
          description: 全名
        phone:
          type: string
          description: 手机号
        role:
          type: string
          enum: [super_admin, admin, manager, staff]
          description: 用户角色
        status:
          type: string
          enum: [active, inactive]
          default: "active"
          description: 用户状态
...
    UserResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/User'
    PaginatedUsersResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/User'
                pagination:
                  $ref: '#/components/schemas/Pagination'

    # 门店相关
...
    Order:
      type: object
      properties:
        id:
          type: integer
          description: 订单ID
          example: 1
        orderNo:
          type: string
          description: 订单号
          example: "ORD202401010001"
        storeId:
          type: integer
          description: 门店ID
          example: 1
        memberId:
          type: integer
          description: 会员ID
          example: 1
        tableNumber:
          type: string
          description: 桌号
          example: "A01"
        orderType:
          type: string
          enum: [dine_in, takeaway, delivery]
          description: 订单类型
          example: "dine_in"
        subtotal:
          type: number
          description: 小计
          example: 156.00
        totalAmount:
          type: number
          description: 总金额
          example: 156.00
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, serving, completed, cancelled, refunded]
          description: 订单状态
          example: "confirmed"
        paymentStatus:
          type: string
          enum: [unpaid, paid, refunded, partial_refunded]
          description: 支付状态
          example: "unpaid"
        items:
          type: array
          description: 订单明细
          items:
            $ref: '#/components/schemas/OrderItem'
        createdAt:
          type: string
          format: date-time
          description: 创建时间
    OrderItem:
      type: object
      properties:
        id:
          type: integer
          description: 明细ID
          example: 1
        itemId:
          type: integer
          description: 菜品ID
          example: 1
        itemName:
          type: string
          description: 菜品名称
          example: "宫保鸡丁"
        quantity:
          type: integer
          description: 数量
          example: 2
        unitPrice:
          type: number
          description: 单价
          example: 38.00
        totalPrice:
          type: number
          description: 小计
          example: 76.00
        specialRequests:
          type: string
          description: 特殊要求
          example: "不要辣"
    CreateOrderRequest:
      type: object
      required: [storeId, items]
      properties:
        storeId:
          type: integer
          description: 门店ID
          example: 1
        memberId:
          type: integer
          description: 会员ID
          example: 1
        tableNumber:
          type: string
          description: 桌号
          example: "A01"
        orderType:
          type: string
          enum: [dine_in, takeaway, delivery]
          description: 订单类型
          default: "dine_in"
        peopleCount:
          type: integer
          description: 用餐人数
          example: 2
        items:
          type: array
          description: 订单明细
          items:
            type: object
            required: [itemId, quantity]
            properties:
              itemId:
                type: integer
                description: 菜品ID
                example: 1
              quantity:
                type: integer
                description: 数量
                example: 2
              specId:
                type: integer
                description: 规格ID
              options:
                type: array
                description: 选项
                items:
                  type: object
                  properties:
                    optionId:
                      type: integer
                      description: 选项ID
                    quantity:
                      type: integer
                      description: 数量
        specialRequests:
          type: string
          description: 特殊要求
          example: "不要辣"
    PaginatedOrdersResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/Order'
                pagination:
                  $ref: '#/components/schemas/Pagination'

    # 会员相关
...
  responses:
    Unauthorized:
      description: 未授权访问
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                example:
                  success: false
                  error:
                    code: "AUTHENTICATION_ERROR"
                    message: "认证失败，请重新登录"
                    statusCode: 401
                    timestamp: "2024-01-01T12:00:00.000Z"
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                example:
                  success: false
                  error:
                    code: "AUTHORIZATION_ERROR"
                    message: "权限不足，无法访问该资源"
                    statusCode: 403
                    timestamp: "2024-01-01T12:00:00.000Z"
...
    ValidationError:
      description: 请求参数验证失败
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                example:
                  success: false
                  error:
                    code: "VALIDATION_ERROR"
                    message: "请求参数验证失败"
                    statusCode: 422
                    details:
                      - field: "email"
                        message: "邮箱格式不正确"
                        value: "invalid-email"
                    timestamp: "2024-01-01T12:00:00.000Z"
...
security:
  - bearerAuth: []